+++ cmd/precise-code-intel-worker/internal/correlation/correlate_test.go
@@ -11,7 +11,7 @@ import (
 )
 
 func TestCorrelate(t *testing.T) {
-	input, err := ioutil.ReadFile("../../testdata/dump.lsif")
+	input, err := ioutil.ReadFile("../../testdata/dump1.lsif")
 	if err != nil {
 		t.Fatalf("unexpected error reading test file: %s", err)
 	}
@@ -24,11 +24,11 @@ func TestCorrelate(t *testing.T) {
 	expectedState := &State{
 		LSIFVersion: "0.4.3",
 		ProjectRoot: "file:///test/root",
-		DocumentData: map[string]lsif.DocumentData{
+		DocumentData: map[string]lsif.Document{
 			"02": {URI: "foo.go", Contains: datastructures.IDSet{"04": {}, "05": {}, "06": {}}},
 			"03": {URI: "bar.go", Contains: datastructures.IDSet{"07": {}, "08": {}, "09": {}}},
 		},
-		RangeData: map[string]lsif.RangeData{
+		RangeData: map[string]lsif.Range{
 			"04": {
 				StartLine:          1,
 				StartCharacter:     2,
@@ -78,7 +78,7 @@ func TestCorrelate(t *testing.T) {
 				MonikerIDs:     datastructures.IDSet{"19": {}},
 			},
 		},
-		ResultSetData: map[string]lsif.ResultSetData{
+		ResultSetData: map[string]lsif.ResultSet{
 			"10": {
 				DefinitionResultID: "12",
 				ReferenceResultID:  "14",
@@ -101,13 +101,13 @@ func TestCorrelate(t *testing.T) {
 			"16": "```go\ntext A\n```",
 			"17": "```go\ntext B\n```",
 		},
-		MonikerData: map[string]lsif.MonikerData{
+		MonikerData: map[string]lsif.Moniker{
 			"18": {Kind: "import", Scheme: "scheme A", Identifier: "ident A", PackageInformationID: "22"},
 			"19": {Kind: "export", Scheme: "scheme B", Identifier: "ident B", PackageInformationID: "23"},
 			"20": {Kind: "import", Scheme: "scheme C", Identifier: "ident C", PackageInformationID: ""},
 			"21": {Kind: "export", Scheme: "scheme D", Identifier: "ident D", PackageInformationID: ""},
 		},
-		PackageInformationData: map[string]lsif.PackageInformationData{
+		PackageInformationData: map[string]lsif.PackageInformation{
 			"22": {Name: "pkg A", Version: "v0.1.0"},
 			"23": {Name: "pkg B", Version: "v1.2.3"},
 		},
@@ -119,6 +119,78 @@ func TestCorrelate(t *testing.T) {
 		ExportedMonikers:       datastructures.IDSet{"19": {}},
 		LinkedMonikers:         datastructures.DisjointIDSet{"19": {"21": {}}, "21": {"19": {}}},
 		LinkedReferenceResults: datastructures.DisjointIDSet{"14": {"15": {}}, "15": {"14": {}}},
+	}
+
+	if diff := cmp.Diff(expectedState, state); diff != "" {
+		t.Errorf("unexpected state (-want +got):\n%s", diff)
+	}
+}
+
+func TestCorrelateMetaDataRoot(t *testing.T) {
+	input, err := ioutil.ReadFile("../../testdata/dump2.lsif")
+	if err != nil {
+		t.Fatalf("unexpected error reading test file: %s", err)
+	}
+
+	state, err := correlateFromReader(bytes.NewReader(input), "root/")
+	if err != nil {
+		t.Fatalf("unexpected error correlating input: %s", err)
+	}
+
+	expectedState := &State{
+		LSIFVersion: "0.4.3",
+		ProjectRoot: "file:///test/root/",
+		DocumentData: map[string]lsif.Document{
+			"02": {URI: "foo.go", Contains: datastructures.IDSet{}},
+		},
+		RangeData:              map[string]lsif.Range{},
+		ResultSetData:          map[string]lsif.ResultSet{},
+		DefinitionData:         map[string]datastructures.DefaultIDSetMap{},
+		ReferenceData:          map[string]datastructures.DefaultIDSetMap{},
+		HoverData:              map[string]string{},
+		MonikerData:            map[string]lsif.Moniker{},
+		PackageInformationData: map[string]lsif.PackageInformation{},
+		NextData:               map[string]string{},
+		ImportedMonikers:       datastructures.IDSet{},
+		ExportedMonikers:       datastructures.IDSet{},
+		LinkedMonikers:         datastructures.DisjointIDSet{},
+		LinkedReferenceResults: datastructures.DisjointIDSet{},
+	}
+
+	if diff := cmp.Diff(expectedState, state); diff != "" {
+		t.Errorf("unexpected state (-want +got):\n%s", diff)
+	}
+}
+
+func TestCorrelateMetaDataRootX(t *testing.T) {
+	input, err := ioutil.ReadFile("../../testdata/dump3.lsif")
+	if err != nil {
+		t.Fatalf("unexpected error reading test file: %s", err)
+	}
+
+	state, err := correlateFromReader(bytes.NewReader(input), "")
+	if err != nil {
+		t.Fatalf("unexpected error correlating input: %s", err)
+	}
+
+	expectedState := &State{
+		LSIFVersion: "0.4.3",
+		ProjectRoot: "file:///__w/sourcegraph/sourcegraph/shared/",
+		DocumentData: map[string]lsif.Document{
+			"02": {URI: "../node_modules/@types/history/index.d.ts", Contains: datastructures.IDSet{}},
+		},
+		RangeData:              map[string]lsif.Range{},
+		ResultSetData:          map[string]lsif.ResultSet{},
+		DefinitionData:         map[string]datastructures.DefaultIDSetMap{},
+		ReferenceData:          map[string]datastructures.DefaultIDSetMap{},
+		HoverData:              map[string]string{},
+		MonikerData:            map[string]lsif.Moniker{},
+		PackageInformationData: map[string]lsif.PackageInformation{},
+		NextData:               map[string]string{},
+		ImportedMonikers:       datastructures.IDSet{},
+		ExportedMonikers:       datastructures.IDSet{},
+		LinkedMonikers:         datastructures.DisjointIDSet{},
+		LinkedReferenceResults: datastructures.DisjointIDSet{},
 	}
 
 	if diff := cmp.Diff(expectedState, state); diff != "" {