+++ internal/codeintel/db/uploads_test.go
@@ -558,7 +558,7 @@ func TestDequeueSkipsLocked(t *testing.T) {
 	dbtesting.SetupGlobalTestDB(t)
 	db := testDB()
 
-	t1 := time.Now().UTC()
+	t1 := time.Unix(1587396557, 0).UTC()
 	t2 := t1.Add(time.Minute)
 	t3 := t2.Add(time.Minute)
 	insertUploads(