+++ shared/src/api/client/connection.ts
@@ -12,7 +12,6 @@ import { createClientContent } from './api/content'
 import { ClientContext } from './api/context'
 import { ClientExtensions } from './api/extensions'
 import { ClientLanguageFeatures } from './api/languageFeatures'
-import { ClientWorkspace } from './api/workspace'
 import { ClientSearch } from './api/search'
 import { ClientViews } from './api/views'
 import { ClientWindows } from './api/windows'
@@ -134,12 +133,11 @@ export async function createExtensionHostClientConnection(
     )
     const clientSearch = new ClientSearch(services.queryTransformer)
     const clientCommands = new ClientCommands(services.commands)
-    subscription.add(new ClientWorkspace(proxy.workspace, services.workspace))
     subscription.add(new ClientExtensions(proxy.extensions, services.extensions))
 
     const clientContent = createClientContent(services.linkPreviews)
 
-    const [newAPI, sub] = initMainThreadAPI(proxy, platformContext)
+    const [newAPI, sub] = initMainThreadAPI(proxy, platformContext, services.workspace)
 
     subscription.add(sub)